feat:解析切换验证方式报文

+++ src/main/java/com/wws/mysqlclient/enums/AttributeKeys.java
@@ -12,8 +12,8 @@ import java.util.concurrent.atomic.AtomicInteger;
  **/
 public class AttributeKeys {
 
-    public static final AttributeKey<AtomicInteger> SEQUENCE_ID_KEY = AttributeKey.newInstance("sequenceId");
-    public static final AttributeKey<AtomicInteger> STATUS_KEY = AttributeKey.newInstance("status");
+    public static final AttributeKey<Byte> SEQUENCE_ID_KEY = AttributeKey.newInstance("sequenceId");
+    public static final AttributeKey<Byte> STATUS_KEY = AttributeKey.newInstance("status");
     public static final AttributeKey<MysqlConfig> CONFIG_KEY = AttributeKey.newInstance("com/wws/mysqlclient/config");
 
 }
